Split new tests into new file
